refactor: Move WriterWorker to its own class
