+++ vicinae/src/services/files-service/file-indexer/file-indexer.hpp
@@ -1,14 +1,8 @@
 #pragma once
-#include <atomic>
-#include <condition_variable>
-#include <cstdlib>
-#include <filesystem>
-#include <mutex>
 #include <qfilesystemwatcher.h>
 #include <qobject.h>
 #include <qsqlquery.h>
 #include <qthread.h>
-#include "common.hpp"
 #include "services/files-service/abstract-file-indexer.hpp"
 #include "services/files-service/file-indexer/indexer-scanner.hpp"
 #include "services/files-service/file-indexer/home-directory-watcher.hpp"
@@ -17,23 +11,6 @@
 #include <qdatetime.h>
 #include <qsqldatabase.h>
 #include <qtmetamacros.h>
-
-class WriterWorker : public NonCopyable {
-  std::unique_ptr<FileIndexerDatabase> db;
-  std::mutex &batchMutex;
-  std::deque<std::vector<std::filesystem::path>> &batchQueue;
-  std::condition_variable &m_batchCv;
-  std::atomic<bool> m_alive = true;
-
-  void batchWrite(const std::vector<std::filesystem::path> &paths);
-
-public:
-  void run();
-  void stop();
-
-  WriterWorker(std::mutex &batchMutex, std::deque<std::vector<std::filesystem::path>> &batchQueue,
-               std::condition_variable &batchCv);
-};
 
 /**
  * Generic file indexer that should be usable in most linux environments.
